Possible solution to a confusing explanation
I was confused by the explanation of grid-area as being from this row to that row, this column to that column. Once I realized it was using coordinates, it all made sense, so I propose providing that as an alternative explanation, or possibly removing the row-to-row, column-to-column explanation. I left it in, since I realize that explanation may work for some. This might also require going back to the introduction of the grid lines and providing the alternative "coordinates" explanation there as well.